backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.4.13=h06a4308_1
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.5=py37h261ae71_0
click=7.1.2=pyhd3eb1b0_0
cryptography=3.4.7=py37hd23ed53_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
fastremap=1.11.2=pypi_0
ffmpeg=4.3=hf484d3e_0
fill-voids=2.0.1=pypi_0
flask=1.1.2=pyhd3eb1b0_0
flask-cors=3.0.10=pypi_0
flask-sqlalchemy=2.5.1=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
imageio=2.9.0=pypi_0
intel-openmp=2020.2=254
itsdangerous=1.1.0=py37_0
jinja2=2.11.3=pyhd3eb1b0_0
jpeg=9b=0
kiwisolver=1.3.1=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.14=0
libidn2=2.3.0=h27cfd23_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h2531618_0
markupsafe=1.1.1=py37h14c3975_1
matplotlib=3.4.1=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.7.2=hbbd107a_1
networkx=2.5.1=pypi_0
ninja=1.7.2=0
numpy=1.20.2=pypi_0
olefile=0.46=py37_0
opencv-python=4.5.1.48=pypi_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
pillow=8.2.0=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
pycparser=2.20=py_2
pydicom=2.1.2=pypi_0
pymysql=1.0.2=py37h06a4308_1
pyparsing=2.4.7=pypi_0
python=3.7.10=hdb3f193_0
python-dateutil=2.8.1=pypi_0
pytorch=1.8.1=py3.7_cuda10.2_cudnn7.6.5_0
pywavelets=1.1.1=pypi_0
readline=8.1=h27cfd23_0
scikit-image=0.18.1=pypi_0
scipy=1.6.2=pypi_0
setuptools=52.0.0=py37h06a4308_0
simpleitk=2.0.2=pypi_0
six=1.15.0=py37h06a4308_0
sqlalchemy=1.3.23=py37h27cfd23_0
sqlite=3.35.2=hdfb4753_0
tifffile=2021.4.8=pypi_0
tk=8.6.10=hbc83047_0
torchaudio=0.8.1=py37
torchvision=0.9.1=py37_cu102
typing_extensions=3.7.4.3=pyha847dfd_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yapf=0.31.0=pyhd3eb1b0_0
zlib=1.2.11=0
zstd=1.4.9=haebb681_0
